consolidate chemotaxis and chemotaxis_control experiments
