+++ vivarium/experiments/chemotaxis.py
@@ -1,17 +1,18 @@
 from __future__ import absolute_import, division, print_function
 
 import os
-import sys
-import uuid
-import argparse
 import copy
+import sys
+import argparse
 
-from vivarium.core.emitter import timeseries_from_data
-from vivarium.library.dict_utils import deep_merge
-from vivarium.core.experiment import (
-    generate_state,
-    Experiment
+from arpeggio import (
+    RegExMatch,
+    ParserPython,
+    OneOrMore,
 )
+
+from vivarium.library.dict_utils import deep_merge
+from vivarium.core.emitter import timeseries_from_data
 from vivarium.core.composition import (
     agent_environment_experiment,
     make_agents,
@@ -27,18 +28,20 @@ from vivarium.compartments.chemotaxis_minimal import ChemotaxisMinimal
 from vivarium.compartments.chemotaxis_master import ChemotaxisMaster
 from vivarium.compartments.chemotaxis_flagella import (
     ChemotaxisVariableFlagella,
-    ChemotaxisExpressionFlagella
+    ChemotaxisExpressionFlagella,
 )
 
 # processes
 from vivarium.processes.Vladimirov2008_motor import MotorActivity
 from vivarium.processes.multibody_physics import agent_body_config
+from vivarium.processes.static_field import make_field
+
+# plots
 from vivarium.plots.multibody_physics import (
     plot_temporal_trajectory,
     plot_agent_trajectory,
     plot_motility,
 )
-from vivarium.processes.static_field import make_field
 
 
 # make an agent from a lone MotorActivity process
@@ -49,13 +52,12 @@ MotorActivityAgent = process_in_compartment(
         'internal': ('cell',)
     })
 
-
+# defaults
 DEFAULT_BOUNDS = [1000, 3000]
 DEFAULT_AGENT_LOCATION = [0.5, 0.1]
 DEFAULT_LIGAND_ID = 'MeAsp'
 DEFAULT_INITIAL_LIGAND = 25.0
 DEFAULT_ENVIRONMENT_TYPE = StaticLattice
-
 
 def get_exponential_env_config():
     # field parameters
@@ -128,8 +130,114 @@ DEFAULT_AGENT_CONFIG = {
     'agents_path': ('..', '..', 'agents')
 }
 
+def set_agent_config(config={}):
+    return deep_merge(dict(DEFAULT_AGENT_CONFIG), config)
 
-# run the simulation
+
+# agent types
+agents_library = {
+    'motor': {
+        'name': 'motor',
+        'type': MotorActivityAgent,
+        'config': DEFAULT_AGENT_CONFIG
+    },
+    'minimal': {
+        'name': 'minimal',
+        'type': ChemotaxisMinimal,
+        'config': DEFAULT_AGENT_CONFIG
+    },
+    'variable': {
+        'name': 'variable',
+        'type': ChemotaxisVariableFlagella,
+        'config': DEFAULT_AGENT_CONFIG
+    },
+    'expression': {
+        'name': 'expression',
+        'type': ChemotaxisExpressionFlagella,
+        'config': DEFAULT_AGENT_CONFIG
+    },
+}
+
+def make_agent_config(agent_specs):
+    agent_type = agent_specs[0].value
+    number = int(agent_specs[1].value)
+    config = agents_library[agent_type]
+    config['number'] = number
+    return config
+
+
+# preset experimental configurations
+preset_experiments = {
+    'minimal': {
+        'agents_config': [
+            {
+                'number': 6,
+                'name': 'minimal',
+                'type': ChemotaxisMinimal,
+                'config': DEFAULT_AGENT_CONFIG
+            }
+        ],
+        'environment_config': DEFAULT_ENVIRONMENT_CONFIG,
+        'simulation_settings': {
+            'total_time': 30,
+            'emit_step': 0.1,
+        },
+    },
+    'master': {
+        'agents_config': [
+            {
+                'number': 1,
+                'name': 'master',
+                'type': ChemotaxisMaster,
+                'config': DEFAULT_AGENT_CONFIG
+            }
+        ],
+        'environment_config': DEFAULT_ENVIRONMENT_CONFIG,
+        'simulation_settings': {
+            'total_time': 30,
+            'emit_step': 0.1,
+        },
+    },
+    'variable': {
+        'agents_config': [
+            {
+                'number': 1,
+                'name': '{}_flagella'.format(n_flagella),
+                'type': ChemotaxisVariableFlagella,
+                'config': set_agent_config({'n_flagella': n_flagella})
+            }
+            for n_flagella in [0, 3, 6, 9, 12]
+        ],
+        'environment_config': DEFAULT_ENVIRONMENT_CONFIG,
+        'simulation_settings': {
+            'total_time': 720,
+            'emit_step': 0.1,
+        },
+    },
+    'mixed': {
+        'agents_config': [
+            {
+                'type': ChemotaxisMinimal,
+                'name': 'motor_receptor',
+                'number': 1,
+                'config': DEFAULT_AGENT_CONFIG
+            },
+            {
+                'type': MotorActivityAgent,
+                'name': 'motor',
+                'number': 1,
+                'config': DEFAULT_AGENT_CONFIG
+            }
+        ],
+        'environment_config': DEFAULT_ENVIRONMENT_CONFIG,
+        'simulation_settings': {
+            'total_time': 720,
+            'emit_step': 0.1,
+        },
+    },
+}
+
+
 def run_chemotaxis_experiment(
     agents_config=None,
     environment_config=None,
@@ -181,156 +289,7 @@ def run_chemotaxis_experiment(
     return simulate_experiment(experiment, settings)
 
 
-def run_mixed(out_dir='out'):
-    total_time = 720
-    emit_step = 0.1
-
-    # configure
-    agents_config = [
-        {
-            'type': ChemotaxisMinimal,
-            'name': 'motor_receptor',
-            'number': 1,
-            'config': DEFAULT_AGENT_CONFIG
-        },
-        {
-            'type': MotorActivityAgent,
-            'name': 'motor',
-            'number': 1,
-            'config': DEFAULT_AGENT_CONFIG
-        }
-    ]
-
-    environment_config = {
-        'type': DEFAULT_ENVIRONMENT_TYPE,
-        'config': get_exponential_env_config()
-    }
-
-    simulation_settings = {
-        'total_time': total_time,
-        'emit_step': emit_step
-    }
-
-    # simulate
-    data = run_chemotaxis_experiment(
-        agents_config=agents_config,
-        environment_config=environment_config,
-        simulation_settings=simulation_settings,
-    )
-
-    # plot
-    field_config = environment_config['config']['field']
-    plot_chemotaxis_experiment(data, field_config, out_dir)
-
-
-def run_variable(out_dir='out'):
-    total_time = 720
-    emit_step = 0.1
-
-    flagella_numbers = [0, 3, 6, 9, 12]
-
-    baseline_agent_config = {
-        'number': 1,
-        'type': ChemotaxisVariableFlagella,
-        'config': DEFAULT_AGENT_CONFIG
-    }
-
-    # configure
-    agents_config = []
-    for n_flagella in flagella_numbers:
-        agent_config = copy.deepcopy(baseline_agent_config)
-        agent_config['name'] = '{}_flagella'.format(n_flagella)
-        agent_config['config'].update({'n_flagella': n_flagella})
-        agents_config.append(agent_config)
-
-    environment_config = {
-        'type': DEFAULT_ENVIRONMENT_TYPE,
-        'config': get_exponential_env_config()}
-
-    simulation_settings = {
-        'total_time': total_time,
-        'emit_step': emit_step}
-
-    # simulate
-    data = run_chemotaxis_experiment(
-        agents_config=agents_config,
-        environment_config=environment_config,
-        simulation_settings=simulation_settings,
-    )
-
-    # plot
-    field_config = environment_config['config']['field']
-    plot_chemotaxis_experiment(data, field_config, out_dir)
-
-
-def run_minimal(out_dir='out'):
-    total_time = 30
-    emit_step = 0.1
-
-    # configure
-    agents_config = [
-        {
-            'number': 6,
-            'name': 'minimal',
-            'type': ChemotaxisMinimal,
-            'config': DEFAULT_AGENT_CONFIG,
-        }
-    ]
-
-    simulation_settings = {
-        'total_time': total_time,
-        'emit_step': emit_step}
-
-    data = run_chemotaxis_experiment(
-        agents_config=agents_config,
-        environment_config=DEFAULT_ENVIRONMENT_CONFIG,
-        simulation_settings=simulation_settings
-    )
-
-    # plot
-    field_config = DEFAULT_ENVIRONMENT_CONFIG['config']['field']
-    plot_chemotaxis_experiment(data, field_config, out_dir)
-
-
-def run_master(out_dir='out'):
-    # TODO -- master requires environment for metabolism external
-
-    agent_type = ChemotaxisMaster
-    total_time = 30
-    emit_step = 0.1
-
-    # configure
-    agents_config = [
-        {
-            'number': 1,
-            'name': 'master',
-            'type': agent_type,
-            'config': DEFAULT_AGENT_CONFIG
-        }
-    ]
-
-    environment_config = {
-        'type': DEFAULT_ENVIRONMENT_TYPE,
-        'config': get_exponential_env_config(),
-    }
-
-    simulation_settings = {
-        'total_time': total_time,
-        'emit_step': emit_step,
-    }
-
-    # simulate
-    data = run_chemotaxis_experiment(
-        agents_config=agents_config,
-        environment_config=environment_config,
-        simulation_settings=simulation_settings,
-    )
-
-    # plot
-    field_config = environment_config['config']['field']
-    plot_chemotaxis_experiment(data, field_config, out_dir)
-
-
+# plotting
 def plot_chemotaxis_experiment(
         data,
         field_config,
@@ -365,36 +324,117 @@ def plot_chemotaxis_experiment(
         print('plot_motility failed')
 
 
+# parsing expression grammar for agents
+def agent_type(): return RegExMatch(r'[a-zA-Z0-9.\-\_]+')
+def number(): return RegExMatch(r'[0-9]+')
+def specification(): return agent_type, number
+def rule(): return OneOrMore(specification)
+agent_parser = ParserPython(rule)
+def parse_agents_string(agents_string):
+    all_agents = agent_parser.parse(agents_string)
+    agents_config = []
+    for idx, agent_specs in enumerate(all_agents):
+        agents_config.append(make_agent_config(agent_specs))
+    return agents_config
+
+
 def make_dir(out_dir):
     if not os.path.exists(out_dir):
         os.makedirs(out_dir)
 
+def add_arguments():
+    parser = argparse.ArgumentParser(description='chemotaxis control')
+    parser.add_argument(
+        '--agents', '-a',
+        type=str,
+        nargs='+',
+        default='"minimal 1"',
+        help='A list of agent types and numbers in the format "agent_type1 number1 agent_type2 number2"')
+    parser.add_argument(
+        '--environment', '-e',
+        type=str,
+        default='exponential',
+        help='the environment type ("linear" or "exponential")')
+    parser.add_argument(
+        '--time', '-t',
+        type=int,
+        default=10,
+        help='total simulation time, in seconds')
+    parser.add_argument(
+        '--emit', '-m',
+        type=int,
+        default=1,
+        help='emit interval, in seconds')
+    parser.add_argument(
+        '--experiment', '-x',
+        type=str,
+        default=None,
+        help='preconfigured experiments')
+    return parser.parse_args()
 
-if __name__ == '__main__':
+
+def run_chemotaxis_simulation():
+    """
+    Execute a chemotaxis simulation with any number of chemotactic agent types
+    """
     out_dir = os.path.join(EXPERIMENT_OUT_DIR, 'chemotaxis')
     make_dir(out_dir)
 
-    parser = argparse.ArgumentParser(description='chemotaxis experiments')
-    parser.add_argument('--minimal', '-n', action='store_true', default=False)
-    parser.add_argument('--master', '-m', action='store_true', default=False)
-    parser.add_argument('--variable', '-v', action='store_true', default=False)
-    parser.add_argument('--mixed', '-x', action='store_true', default=False)
-    args = parser.parse_args()
-    no_args = (len(sys.argv) == 1)
+    args = add_arguments()
 
-    if args.minimal or no_args:
-        minimal_out_dir = os.path.join(out_dir, 'minimal')
-        make_dir(minimal_out_dir)
-        run_minimal(minimal_out_dir)
-    elif args.master:
-        master_out_dir = os.path.join(out_dir, 'master')
-        make_dir(master_out_dir)
-        run_master(master_out_dir)
-    elif args.variable:
-        variable_out_dir = os.path.join(out_dir, 'variable')
-        make_dir(variable_out_dir)
-        run_variable(variable_out_dir)
-    elif args.mixed:
-        mixed_out_dir = os.path.join(out_dir, 'mixed')
-        make_dir(mixed_out_dir)
-        run_mixed(mixed_out_dir)
+    if args.experiment:
+        # get a preset experiment
+        # make a directory for this experiment
+        experiment_name = str(args.experiment)
+        control_out_dir = os.path.join(out_dir, experiment_name)
+        make_dir(control_out_dir)
+
+        experiment_config = preset_experiments[experiment_name]
+        agents_config = experiment_config['agents_config']
+        environment_config = experiment_config['environment_config']
+        simulation_settings = experiment_config['simulation_settings']
+
+    else:
+        # make a directory for this experiment
+        experiment_name = '_'.join(args.agents)
+        control_out_dir = os.path.join(out_dir, experiment_name)
+        make_dir(control_out_dir)
+
+        # configure the agents
+        agents_config = []
+        if args.agents:
+            agents_string = ' '.join(args.agents)
+            agents_config = parse_agents_string(agents_string)
+
+        # configure the environment
+        if args.environment == 'linear':
+            env_config = get_linear_env_config()
+        else:
+            env_config = get_exponential_env_config()
+        environment_config = {
+            'type': DEFAULT_ENVIRONMENT_TYPE,
+            'config': env_config,
+        }
+
+        # configure the simulation
+        total_time = args.time
+        emit_step = args.emit
+        simulation_settings = {
+            'total_time': total_time,
+            'emit_step': emit_step,
+        }
+
+    # simulate
+    data = run_chemotaxis_experiment(
+        agents_config=agents_config,
+        environment_config=environment_config,
+        simulation_settings=simulation_settings,
+    )
+
+    # plot
+    field_config = environment_config['config']['field']
+    plot_chemotaxis_experiment(data, field_config, control_out_dir)
+
+
+if __name__ == '__main__':
+    run_chemotaxis_simulation()
